requirements.txt
# Core Web Framework
streamlit>=1.28.0

# AI and Machine Learning
openai>=1.3.0
numpy>=1.24.0

# Image Processing
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# Environment and Configuration
python-dotenv>=1.0.0

# Optional: For enhanced performance
requests>=2.31.0
